seeding is fully functional -- LaughData class moved to new file
